Amang/requirements.txt
pyrogram
tgcrypto
pytz
py-tgcalls==0.9.7
yt-dlp
youtube-search-python
motor
dnspython
GitPython
pykeyboard
bs4
Pillow
pymediainfo
opencv-python-headless
requests
ffmpeg-python
unidecode
apscheduler
attrify
pyromod==1.1.0
html2text
removebg
aiofiles
aiohttp
gpytranslate
gtts
wget
telegraph
heroku3
openai
httpx[http2]
uvloop
httpx
python-dotenv
pytimeparse
